Fix problem of contact's birthday being one day later than selected
If a user selected a birthday or anniversary for a contact, the date
would sometimes be one day earlier or later than what the user
selected.  This would happen if the local time was on a different day
from UTC, which is true in the evening in Pactific Time.  The problem
has been fixed by being more careful about which times are local and
which are UTC.

Fix nachocove/qa#141