For testing purposes, fill the calendar with events
Add a new tab to the main tab controller: Manage Items.  This tab has
buttons to fill the calendar with events and to delete everything from
the calendar.  This will be useful for testing how the app performs at
scale.  The events that are created are all appointments.  Creating
meetings will be a later enhancement.

The Manage Items tab is not present in any of the production versions
of the app.  The code to enable it is in an "#if false" section.  The
tab should only be enabled (by changing "false" to "true") when
running the app in the simulator.