Fix event detail view when there are exactly five attendees
When a meeting had exactly five attendees, the event detail view was
showing all five of them, but the fifth one had a circle with a +1
superimposed over it.  When there is exactly five attendees, we want
to simply show all five.  When there are more than five, then show
just for of them and use the last slot to indicate how many more
attendees are not being shown in this view.

This same bug was fixed a while ago in BodyCalendarView, but not in
EventViewController.