Prevent events that end before they start
Add some validation when the user taps the Send or Done button in the
edit event view.  If the title of the event is empty, or if the ending
time of the event is earlier than the starting time, then tapping the
button doesn't save the event or dismiss the view.  Instead, an alert
pops up explaining the problem and the user is left in the editor.

Fix #1495