Refresh views when the system time zone changes
Add a listener to detect when the system time zone changes.  Change
the event listener for most of the list views to refresh the view when
the time zone changes, so that any times on the views will be
correctly updated.  (The message detail and event detail views will
not update, because those views are not designed to be refreshed.
Adding the code to refresh them is not worth the effort.)

Fix nachocove/qa#196