Correctly display all-day events when device changes time zones
All-day events were being shown one day too early if the user changed
the time zone on his device to a more westerly time zone.  To fix this
problem, change the way that all-day McEvents are stored.  There is
now a boolean flag in McEvent to indicate an all-day event, and the
start and end times for the event are always midnight in UTC, rather
than being midnight in the local time zone.

This change requires a migration.  Rather than having the migration
figure out which McEvents need to have the AllDayEvent flag set,
instead we just set a flag on all-day McCalendar items that will cause
the McEvents to be recreated.

Fix nachocove/qa#203