Create wrapper classes for alert views and action sheets
UIAlertView and UIActionSheet have both been deprecated in iOS 8.
Create wrapper classes for both of those, NcAlertView and
NcActionSheet respectively.  The wrapper classes will use UIAlertView
and UIActionSheet when the app is running on iOS 7, and they will use
UIActionController, the replacement for both UIAlertView and
UIActionSheet, when running on iOS 8.  The wrapper classes present a
simplified interface and do not have the full functionality of the
underlying classes.

All uses of UIActionSheet were changed to use NcActionSheet.  Most
uses of UIAlertView were changed to NcAlertView, but some uses of
UIAlertView were left for now.