Improve log messages about long running transactions
When measuring the duration of RunInTransaction, separate the time
spent waiting to acquire the database write lock from the time spent
doing the actual work.  When reporting a long transaction, use much
lower threshholds for the UI thread than for other threads, because a
long transaction on the UI thread is a more serious problem than a
long transaction on a background thread.