Allow Attend/Maybe/Decline one occurrence of a series
When the user responds (Attend/Maybe/Decline) to a recurring meeting
from the event detail view, prompt the user as to whether the response
is for the entire series or just the one occurrence.

When the user declines a meeting from the event detail view, prompt
the user to confirm the action, since declining the meeting also
removes it from the calendar in a way that is hard to undo.

(When the user declines a recurring meeting, the two prompts are
combined into screen, so the user is prompted at most once.)

Clean up some of the code in EventViewController, removing commented
out code and changing function names.

Fix nachocove/qa#67