Improve the handling of event descriptions and attachments
When the user edits an event, the app is smarter about extracting
plain text from non-plain text formats when it fills in the
description field of the editor.  Previously, the description field
could be blank or full of HTML tags when the user edited the event.
Now the description field in the editor should have the plain text
version of the description.

If the user edits the event but doesn't change the description in any
way, then the <Body> field, which is used to hold the description, is
left out of the application data when sending the update to the
server.  The server interprets this as "don't change the body" rather
than "delete the body."  This change allows the user to edit the event
without messing up a nicely formatted event description.

The handling of event attachments has been improved, and attachments
are now handled as well as they can be given the limitations of the
various servers.  (Except for one known bug: Attachments are not quite
correct when a recurring meeting has exceptions. That will be a future
update.)

Fix nachocove/qa#82
Fix nachocove/qa#30
Fix nachocove/qa#213