Improve the display of most event descriptions
Make some improvements to how BodyView decides which parts of MIME
messages to display and how RTF text is displayed.  There are two main
changes:

1. Use UIWebView instead of UITextView to display RTF.  UIWebView does
a better job of rendering the more complicated formatting such as
tables.  It is still not as nice as if the data were originally in
HTML, but it is better than it was before.  This will improve the
display of many event descriptions, especially meetings that
originated from a GFE account.

2. Use the item's NativeBodyType field when deciding which part of a
multipart/alternative MIME section to display.  This should have
little effect on the display of e-mail messages, but it will result in
events with plain text descriptions being displayed as plain text
rather than as RTF.