Links in plain text messages are now clickable
BodyView now uses UIWebView instead of UITextView to display plain
text.  This avoids a bug in UITextView that was hard to work around,
and it standardizes on UIWebView for displaying all text.  The tricky
part was figuring out how to specify the font that UIWebView uses to
display plain text.

Fix #1184