Allow user to cancel one instance of a recurring meeting
Allow the user to cancel one instance of a recurring meeting where the
user is the organizer of the meeting.  A cancellation notice will be
sent out to all the attendees.  The cancel button appears at the
bottom of the event detail view (but only when the event is an
occurrence of a recurring meeting where the user is the organizer).
The user is prompted to confirm the operation before the cancellation
happens.

Fix nachocove/qa#68