Mark meeting as canceled when processing a meeting cancelation message
Most servers change a meeting's status to canceled when the server
processes an incoming meeting cancelation message.  But some servers
don't do that in all cases.  So change the app to do it.  When the app
processes an incoming meeting cancelation message, it will mark the
meeting as canceled.  The change is only made locally; it is not
syched to the server.

This was tested with servers that mark the meeting as canceled,
servers that don't (Exchange 2007), and servers that delete the
meeting as soon as the cancelation notice arrives (Google).  It was
tested with regular meetings, recurring meetings, and individual
occurrences of recurring meetings.

Fix #1420