Fix bogus download failure message for empty event descriptions
When an event with an empty description was created within the app,
the McCalendar item wouldn't have any McBody at all.  This would
result in a "Download failed. Tap here to retry." message when
displaying the event details.

Tweak EditEventViewController so that a McBody is always created for a
new event.

Fix nachocove/qa#312