Avoid "Download failure" messages for events with empty bodies
Exchange 2007 servers will sometimes send a Body element with no Data
element for calendar events that have an empty description. The app
was interpreting this as if the body had not yet been downloaded. This
resulted in a "Download failure" message in the event detail view
because any attempt to download calendar bodies will fail.

Change the code that parses Body elements to notice this situation and
treat it as if the Data element were empty rather than missing.

Fix #323