Don't process meeting cancelation notices that are owned by the user
To work around a server bug, ignore meeting cancelation messages where
the current user is the organizer of the meeting.  (See the big
comment in the code for more details.)

Get the subject of outgoing meeting cancelation messages correct when
canceling a single occurrence of a recurring meeting where the
occurrence is an exception with a different title.