Don't crash when editing an event that has been deleted
If an event is deleted while the event's detail view is open, and the
user then taps the edit button, the app was crashing because the event
being edited was no longer in the database.

Change the code to detect the situation, show an alert, and pop out of
the editor back to the event detail view.

Fix nachocove/qa#321