"Remove from calendar" removes single occurrence when appropriate
When the user receives a meeting cancelation message for a single
occurrence of a recurring meeting and taps the "Remove from calendar"
button in the message detail view, remove only that occurrence of the
meeting.  Previously, the entire series was being removed.

Fix the code for the "Remove from Calendar" button in the event detail
view to correctly figure out whether to remove the individual
occurrence or the entire series.

When the user receives a meeting request for a singel occurrence of a
recurring meeting, don't try to figure out whether or not the user has
already responded to the request.  Always show the
Attend/Maybe/Decline buttons unselected in this case.

When creating the ActiveSync XML for a calendar item exception,
include only the Deleted and ExceptionStartTime fields for exceptions
that represent deleted occurrences.  In this situation, all the other
fields are meaningless.

When creating the ActiveSync XML for a calendar item, ignore the
AppointmentReplyTime field.  That field is for sync responses only and
should never be in sync requests.

Fix a typo in AsMeetingResponseCommand.ProcessResponse that could
result in a NullReferenceException.

Fix nachocove/qa#307