Cleanup class EditEventViewController
Convert EditEventViewController to use NcUIViewControllerNoLeaks,
which also involved changes to classes UcAttachmentBlock and
UcAttachmentCell.  Unfortunately, this has not stopped the class from
leaking memory, so nachocove/qa#295 can't be considered fixed yet.  I
haven't figured out yet where the leaks are coming from.  But these
changes are a necessary step toward fixing the leaks, so the changes
are being committed.

Fix the behavior of the description field when it is being edited.
The view no longer scrolls unexpectedly, the cursor is always kept
on the visible part of the screen, and there is no useless white
space.
Fix nachocove/qa#257

Fix the animation when expanding and collapsing the start date picker
and end date picker.

Clean up the code (beyond what was necessary to avoid memory leaks).
Unused code was removed.  Code for the phone number field, which had
been commented out, was removed because it was getting out of date.