Fix the default event start/end times for events late in the day
Rewrite CalendarHelper.DefaultMeeting(), being more careful about the
default start and end times for the event.  This fixes two bugs:

Creating an event when the local time was between 10:30pm and 11:30pm
would result in the default end time being 23 hours earlier than the
start time instead of one hour later.

Creating an event when UTC is a different day than local time (which
is true in the late afternoon and evening for Pacific Time) would
result in the event being a day later than it should be.

Fix nachocove/qa#362