Fix the encoding of WBXML opaque data
(None of the code being changed is currently used by the app, which is
why the bugs in the code have not been noticed.  These changes do not
fix any current bugs, and they have not been thoroughly tested.  But
the code might be used in the future, for example if the app uses the
ItemOperations/Move command, so it is good to fix the bugs now.)

For WBXML opaque fields where the app doesn't need to interpret the
data (only use it as an identifier), the app uses a base-64 encoded
string of the data internally (because a string is easier to manage
than a sequence of bytes).  Decoding the bytes and creating the
base-64 string was correct.  But the code to encode a base-64 string
into the correct byte sequence was completely broken.  This has been
fixed.

When encoding opaque data that is stored in a file, transfer the bytes
of the file directly into the WBXML data, rather than reading the file
in as a string and then converting the string back to a sequence of
bytes.  (There are two versions of this code.  The version in
EmitNode() is used for the body of an outgoing e-mail message, and
that one was already correct.  This update fixes the version in
EncodeNode(), which is not used.)

When encoding opaque data that is stored in a non-base-64 string, use
UTF-8 encoding (which preserves all of the information in the string)
rather than ASCII encoding (which loses some information).  This code
should never be run, since all opaque fields are stored as base-64
strings or in a file.  But the code needs to be here for
completeness.