Remove some useless information from message previews
Remove the following strings from message previews that appear in
message list views:
	\[http.*\]
	\[cid.*\]
	\[image.*\]
	<http.*>
These strings are usually less useful to the user that the other
information around them, so it is best to not display them in the very
small message preview field.

This update improves nachocove/qa#3, but it does not fully resolve the
issue.  More work needs to be done to close the issue.