Correctly handle RTF bodies
Fix the app to correctly handle RTF item bodies.  (This does not
affect the handling of MIME messages with a "text/rtf" part.  That
code was already correct.  This fixes the case where the ActiveSync
body type is RTF.  The sync options that the app uses means that the
app should never receive an RTF body.  But in case the server ignores
the options and sends an RTF body, the app should do the right thing.)

The server sends RTF bodies in a base-64 compressed format.  The app
was expecting normal RTF.  Fix the app to convert the base-64
compressed RTF into normal RTF before attempting to display the RTF.

Fix nachocove/qa#375