Improve the algorithm for stripping junk from message previews
The algorithm for adjusting message preview text has been improved.
See the code for details.

To have useful text left over after stripping out the junk, the amount
of data initially fetched had to be increased from 255 bytes to 500
bytes (which is the same amount of data that iOS Mail and Outlook
Mobile fetch for their previews).  But the <Preview> element in
ActiveSync is limited to 255 bytes, so the app had to be changed to
not use <Preview> at all instead ask for 500 bytes of the plain text
version of the body.

Fix nachocove/qa#3