Increase preview fetch size from 255 bytes to 500 bytes
Now that the app strips junk from the message preview text, 255 bytes
is not big enough to consistently have enough useful text left over
after stripping away the junk.  So increase the preview fetch size to
500 bytes.  (That is still not big enough for all messages, but it is
big enough for most, and it is the size used by iOS Mail and Outlook
Mobile.)  Unfortunately the <Preview> sync option is limited by the
spec to 255 bytes.  So the app has to stop using <Preview> and instead
ask for a plain text body truncated to 500 bytes.