Create UIDatePickers only when they are needed in EditEventVC
UIDatePicker leaks memory.  Quite badly.  A couple megabytes every
time one is added to the view hierarchy.  I haven't found a way to
avoid this leak.  So EditEventViewController has been changed to
create its UIDatePickers only when the user expands the start or end
time fields to expose the date picker.  This saves a few megabytes
whenever the user edits an event without adjusting the start or end
times.

This helps with nachocove/qa#295, but it does not resolve the issue.