Don't check MIME content type when resolving a cid: URL
When resolving a cid: URL (which is used to reference an image that is
in another MIME entity in the same message body), don't check the
content type of the MIME entity with the image.  We have seen a
message where the content type for the image was wrong.  The app
should be permissive when parsing inputs and should be able to display
the image anyway.

Fix https://support.nachocove.com/helpdesk/tickets/99