Move strategy code for body and attachment download to AsStrategy
The code that decided whether or not to download attachments when
downloading a message body was in AsItemOperations.  That code has
been moved to AsStrategy and improved.

Improve the code in AsStrategy that figures out speculative downloads
of message bodies and attachments.  The code now takes into account
the quality of the network connection when deciding how much to
download.