Don't crash when a McCalendar has a null TimeZone field
We encountered a crash which seemed to be caused by a McCalendar item
with a null TimeZone field.  It is not clear how the app got into that
state, but since the TimeZone field normally comes from the server
(i.e. it's an external input) the app shouldn't crash in this
situation.  Change the AsTimeZone constructor to gracefully handle a
missing or otherwise malformed encoded string.  When this happens, it
outputs an error and uses UTC.

Fix nachocove/qa#419