Fix several minor issues with the attendee list view
The user is allowed to toggle the type of an attendee between required
and optional without having to edit the meeting first.  But doing this
had no effect; the change was lost.  Fix the app so this gesture does
what it is supposed to do.

Don't crash when the user taps "Send Invite" when viewing an exception
of a recurring meeting.

Don't crash when the user changes the attendee type twice in a row
when viewing an exception of a recurring meeting.  This was fixed by
disabling the "Required"/"Optional" button when viewing a recurring
meeting.  Users are currently not allowed to edit a recurring meeting,
so they shouldn't be allowed to indirectly edit it by changing the
attendee type.

Don't crash when the user deletes a meeting from another client while
the attendee list view is visible, and the proceeds to change the
attendee type twice in a row.

Fix the behavior when a meeting has an attendee with the type
"Resource".  (Such an attendee can't be created in the app, but it can
be created in some other clients.)  Previously, the app was showing
the Required button for a Resource attendee, but was changing the type
to Optional when the button was pressed.

Fix nachocove/qa#408