Fix bug of sending meeting requests with RTF code in description
When a meeting was initially created on a different client with a
formatted body, then editing the meeting in the app would result in
the meeting request message being sent out with the raw RTF as the
meeting's description.  This would also happen when choosing "Send
Invite" from the attendee view to resend an invitation.

Fix the problem by converting a formated meeting description into
plain text before sending out a meeting request.

Fix nachocove/qa#406