Download bodies of messages that are visible in a message list view
Whenever a message becomes visible in a message list view, go ahead
and download its body.  Don't wait for the user to open the message
detail.  This avoids many of the situations where the user sees a
spinner for a second or two in the message detail view.

Fix nachocove/qa#420