Don't crash when creating an appointment for a non-Exchange account
Change the way that EditEventViewController figures out the McAccount
to use.  It no longer assumes that there is a single Exchange account.
This avoids a crash when creating an appointment when there are no
Exchange accounts on the device.  There are still problems with the
created appointment, and attempting to create a meeting still causes a
crash.  But with this update there is one less path that triggers a
crash.

Fix nachocove/qa#442