Fix a couple of crashes in the device contact and calendar code
The code that cleans up deleted device contacts and device calendar
items was not protecting against the case where the device contacts or
device calendars could not be accessed, which could result in a
NullReferenceException.

When a pair of tasks for synching device contacts and device calendar
items overlapped, a NullReferenceException could result.  Add checks
and locking to prevent this from happening.