Fix the device protocol sync state machine
When a device contants/calendars sync is interrupted by an abate
signal, wait until the sync has stopped before restarting the sync.
This was done by adding a pair of Cancelling states to the state
machine.

Make some other fixes to the states and transitions in the device
protocol state machine.