Event detail fixes: isOrganizer calculation and editable calculation
Three changes to EventViewController:

An event cannot be edited if it came from the device calendar.

Don't display "You are the organizer" for appointments.  Leave that
section out entirely, since that statement is redundant and not useful
for appointments.

Fix the way that isOrganizer is calculated.  The old code for
isOrganizer was getting the wrong answer for some events in a Hotmail
or Outlook.com account.

Fix nachocove/qa#278