DeviceProtoControl: Don't crash on AbateOn/AbateOff/AbateOn
If the DeviceProtoControl state machine was in state SyncW and then
received AbateOn, AbateOff, and AbateOn events without seeing a
SyncCancelled event, the app would crash because the cancellation
token would be unexpectedly null when processing the second AbateOn.
Fix the code to not crash when the cancellation token is null, since
this situation is normal.