Don't try to ignore certain device calendars during sync
Remove the code that tries to identify device calendars that should be
ignored because the app is already synching those calendars.  The
algorithm being used was unreliable, and only worked for Gmail.  We
need to come up with a different approach for identifying device
events that are duplicates (which will probably be based on individual
events rather than folders).