Avoid duplicate events in the calendar view
When identifying events to show in the calendar view, eliminate
duplicates that can arise when the same event is in several different
calendars that are tracked by the app.  The calendar item's UID is
used to identify duplicates.  When choosing the event to show, choose
the current account over other accounts, and any account over the
device account.