Dots in calendar view rows are colored to show the containing folder
The dots on the left of each entry in the calendar view are now given
a color to match the folder that owns the event.  This provides the
user with a way to know where the event came from without having to
open the event detail.  (This does not affect the dots under the date
in the week view or the month view, just the ones in the rows for
individual events.)

Make the same change for the dot on the left of the hot event view.