Don't crash when contact detail view is opened with no contact
When viewing a meeting that came from the device calendar, the
organizer and attendees don't have any corresponding McContacts.  So
opening the contact detail view results in a
ContactDetailViewController with a null contact.  The contact detail
view code did not handle that situation.  The app would crash when
loading the view, and when that was fixed, the app would crash when
the user interacted with the view in any way.

ContactDetailViewController has been fixed to not crash when it is
given a null contact.

(In a later update, I might change EventViewController to not open the
contact detail view at all if there is no McContact, but I want to
make the contact detail view more robust first.)