Don't crash when message is deleted while message detail view is open
Some recent telemetry changes assume that the message is still
available in MessageViewController.ViewWillDisappear().  That is not
true if the message was deleted while the view was open.  One common
way for the message to be deleted is for the user to accept a meeting
request.

Change the code to protect against a missing message.