Improve the algorithm for choosing the hot event
The hot event view would often show an event that was not the event
that the user was most interested in seeing.  For example, it would
always show an all-day event for the day instead of events that happen
during the day.

Change the algorithm for choosing the hot event to take a number of
factors into account.  As a couple of examples, events that start in
the next 30 minutes are preferred over events that are currently in
progress, and non-all-day events are usually preferred over all-day
events.