Missing '!' was causing crash when adding event to Gmail calendar
A mistake during refactoring resulted in a crash when the user tried
to create a calendar item when the current account was a Gmail or IMAP
account.  Fix the problem so that the user gets a proper error message
in this situation.

Fix nachocove/qa#520
Fix https://support.nachocove.com/helpdesk/tickets/156