Do better at identifying service-type from the email address
Don’t use Account.AccountService, since that’s not reliable. It’s a UI
construct, and at best can be a hint for us. A user can use the
‘iCloud’ button (service type will be iCloud) or the ‘IMAP’ button
(service type will be IMAP_SMTP), and we need to make sure to still do
the right thing.
Do some string compares in ImapDiscoveryCommand.