Fix blank entries in calendar view. Several device calendar fixes.
Fix nachocove/qa#160  Fix the problem of blank entries in the calendar
view.  The bug was caused by incorrect use of the Hidden property for
views.  When displaying an event that was recently deleted (but hadn't
been removed from the view controller's table yet), the main view was
set to Hidden, which turned off it and all its children.  When the
cell was later reused for a real event, "Hidden = false" was done for
all the child views, but never for the main view.  So the cell
remained blank forever.

When a device calendar item was deleted, the app's McCalendar item was
correctly deleted.  But the McEvent for that McCalendar was never
deleted from the database.  This resulted in a blank entry in the
calendar view.  Fix the device calendar sync code delete McEvents when
deleting McCalendars.  (And other ancillary data for the McCalendar,
such as McRecurrence and McAttendee.  These were also left behind, but
those orphans were not user visible.)

If a device calendar item was deleted while the app was synching the
device calendar, the app would see a device calendar item with empty
and inconsistent fields.  Fix the device calendar sync code to
recognize this situation and handle it gracefully.

When the only activity that happens when synching the device calendar
is to delete calendar items, make sure an EventSetChanged event gets
fired.