Only sync one year's worth of device calendar events
Sync just one year of device calendar events, rather than synching
everything that iOS will give us.  Asking for all events was resulting
in the app doing a lot of work that was of no real use to the user.

Fix on oversight in the conversion from iOS's EKRecurrenceRule to the
app's McRecurrence, which was resulting in an ominous sounding (but
harmless) warning message.

Fix nachocove/qa#629