Don't resync device contacts whenever device cal sync was interrupted
Whenever the synching of the device calendar was interrupted (which
happens with every abate signal), the device contacts would be
resynched from the beginning once the abate-off signal was received.
Fix the device account sync code so that the sync will resume the
calendar sync from where it left off and not redo the device contacts.

Fix nachocove/qa#610