Stop the device contacts/calendar sync when app hibernates
When the app has finished its background processing and needs to go
into hibernation, the task to sync the device contacts and calendar
was not shutting down.  It would continue across the app shutdown.
Fix the problem by having the task listen to both its own internal
cancellation token and the general cancellation token from NcTask.