Use less memory when constructing MIME; fix image resize bug
When sending an outgoing message that has attachments, construct the
MIME for the message without loading the attachments into memory.  The
MIME is created by streaming the attachments from one file to another.
This was done by copying and modifying two MimeKit utility classes.
This is an intermediate fix and is only applicable to sending a
message or saving a draft.  A better fix, which could be used to
reduce memory in other places where MimeKit is used, requires changing
some classes closer to the core of MimeKit.  That fix may (or may not)
be attempted later.

Fix nachocove/qa#631.  When sending a message with large image
attachments, the "Medium" and "Large" choices now do what they say
they do.