Fix incorrect timeout values for ActiveSync SendMail command
Fix two bugs in the ActiveSync command timeout calculations, both of
which affected primarily the SendMail command.

When an operation is retried, the timeout value is increased.  A bug
in the way this increase is calculated (using TimeSpan.Seconds instead
of TimeSpan.TotalSeconds) was resulting a timeout that was much too
short if the original timeout had been more than a minute.

The SendMail command calculates an appropriate timeout value based on
the size of the outgoing message.  But the timeout value was not being
set until after the timeout timer had been created.  So the initial
attempt of the SendMail command was always using the default timeout,
not its custom timeout.

Restructure the code for the timeout value for ActiveSync HTTP
operations to avoid these two problems.