Better handling of device event with unknown time zone
Do a better job of handling the situation where an event from the
device calendar has a time zone that cannot be looked up.  Instead of
logging an error with an exception stack trace and ignoring the event,
log a warning and continue processing the event in the local time
zone.

Fix nachocove/qa#698