Don't notify about deleted messages; avoid duplicate notifications
nachocove/qa#722  This is an improvement, but not a complete fix.
Notifications about deleted messages should be fixed.  Duplicate
notifications should be greatly reduced, but not completely
eliminated.

The "Inbox" setting for notifications was looking for messages in any
folder, not just the Inbox.  Change the code to only look in the
Inbox.

When deciding if a message needs a notification, ignore any message
that is in the deleted items folder, even if it is hot or from a VIP.

Keep track of the message IDs of messages for which a notification has
been posted.  Don't notify the user about a message with a message ID
that is in this set of IDs.  This works around the issue of Gmail
placing the same message in multiple folders.