Do a better job with Pull-to-sync.
Don’t resolve the pending after the OpenOnly. Defer it until after the
actual sync.
Combine the ‘SyncDone’ code into one function for sanity.
Remove PrioSyncFolder, since it’s not needed anymore. It was a hack to
match up the OpenOnly and the Sync, which is now done via the deferred
pending.