fix nachocove/qa#762 - crash when gesturing in swipe view
A crash was reproducible by swiping the view with one finger and
then tapping outside the view with another finger while still
swiping with the first.  The second finger's tap would set swipingView
to null, but the pan gesture callback would continue to be called since the
first finger hadn't let up.

Fixed by having the tap callbacks do nothing if a pan gesture is active.