refresh the new messageSource after switching accounts
In order to preserve the behavior that existed before, when MaybeRefreshThreads
was called (and in turn called messageSource.RefreshEmailMessages), we'll do the
refresh in SwitchToAccount.

Turns out that some sources, like NachoDeferredEmailMessages call Refresh
in their constructor, making this a duplicate call.  But others like
NachoEmailMessages don't call Refresh themselves, making this a required call.