Keep current the number in the scroll-to-today button on calendar view
The scroll-to-today button in the navigation bar of the calendar view
contains the day of the month.  That number was being set when the
view was created, but was never adjusted later.  It would become wrong
if the user kept the app running for more than a day.

Update the number if necessary whenever the calendar view becomes
visible or when the app is brought to the foreground.

The number will not be adjusted when the user is looking at the
calendar view when the clock strikes midnight.  But it will be
corrected the next time the moves away from the calendar view and then
back to it.

Fix nachocove/qa#809