Update the hot event when the app returns to the foreground
If the user backgrounded the app while in the Hot view, the hot event
was not being updated when the user brought the app to the foreground.
This could result in the Hot view showing an old event.

Fix this by having HotEventView listen for foreground events and then
updating the hot event.