adjust switch-account segue image capture to align properly
fix nachocove/qa#863

The previous code would fail to properly position the captured image
if the source view was scrolled.

What we want is to offset the captured image by the distance from the
outermost view to the top of the source view.  The source view frame
is in the coordinates of the source view's superview.  So, we should
convert the frame from the source view's superview to the outermost view.