Don't call MonitorReport when an NcTask's start is delayed
When the start of a new NcTask was delayed more than 200 ms,
MonitorReport was called to dump the app's current status.  When the
app is busy, running MonitorReport can take a couple of seconds.  This
delays the start of the new task exactly when the app cannot afford
any more delays.

Remove that call to MonitorReport.  Tweak the contents of the report
slightly.