Lower the timeout value for LoadBytes from 3 min to 20 or 60 sec
The timeout value for LoadBytes was 3 minutes.  That was a lot longer
than any well behaved LoadBytes call should take.  The timeout has
been lowered to 20 seconds.  Or 60 seconds if the HTTP response
headers claim that the response is more than 100,000 bytes.

The new timeout value might be a little low.  We will have to watch
the logs and see how often it times out.

nachocove/qa#826