Stop the back end first when the PerformFetch timer fires
When the PerformFetch timer fires, the app has about five seconds left
to finish its work before it can be killed by iOS.  We have seen some
cases where the app is so busy that the shutdown work can't finish
within those five seconds.

Change the timer code to stop the back end as soon as the timer fires.
The hope is that this will free up the CPU enough so that all the
other shutdown work can finish within five seconds.

Add more information to the log messages that report the progress of
the PerformFetch, so it is easier to track what is happening.

This is the first part of nachocove/qa#938.  There is more work to be
done before it can be resolved.