PerformFetch shouldn't wait for PushAssist when PushAssist is disabled
When the app fails to register for remote notifications, the
PushAssist code is disabled.  But the PerformFetch code was still
waiting for the PushAssist code to do its thing.  This resulted in the
PerformFetch code timing out rather than completing normally when it
should have.

Fix the PerformFetch code to recognize when the PushAssist code is
disabled.

This is part 2 of nachocove/qa#938.  There is still more to come.