ImapCommand.Cancel() and ImapDisconnectCommand should not block
When ImapCommand.Cancel() is called during back end shutdown, it can't
block at all waiting for the command to process the cancellation.  It
has to issue the cancellation and then return.

ImapDisconnectCommand is usually executed during back end shutdown.
So it can't block while waiting to get the lock.  If the lock is not
available, start a separate task to do the disconnect.

nachocove/qa#938 part 3, with at least one more change yet to come.