Avoid race condition that would result in a stopped back end
An earlier update that called BackEnd.Stop() as soon as the
PerformFetch timer fired introduced a race condition that could result
in the back end being stopped while the app was in the foreground.  If
WillEnterForeground() was called by iOS after the PerformFetch timer
fired but before CompletePerformFetch() was called, the app would move
to the foreground state without BackEnd.Start() ever being called.

To avoid this race condition, the earlier update was mostly undone.
BackEnd.Stop() is still called as soon as possible, but "as soon as
possible" is not until control has been transferred to the UI thread
via a status indicator event.

nachocove/qa#938 part 5, and most likely the last change for that issue