When hitting reply, be smart about not replying to self
fixes nachocove/qa#606

The issue talks about replying from the sent folder, but I chose to
key off the From address instead of where the message is located.
Basically, if you reply to message sent from the selected account,
the compose view is populated with the original messages's To list.
This is consistent with behavior in other clients.